backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
bcrypt
PyJWT
pydantic
pydantic[email]
google-genai
